move text utility functions into textUtilities
