+++ src/compiler/roact.ts
@@ -2,8 +2,9 @@ import * as ts from "ts-morph";
 import { compileExpression } from ".";
 import { CompilerState } from "../CompilerState";
 import { CompilerError, CompilerErrorType } from "../errors/CompilerError";
+import { bold, suggest } from "../textUtilities";
 import { getType, isArrayType } from "../typeUtilities";
-import { bold, skipNodesDownwards, suggest } from "../utility";
+import { skipNodesDownwards } from "../utility";
 
 const ROACT_ELEMENT_TYPE = "Roact.Element";
 export const ROACT_COMPONENT_TYPE = "Roact.Component";
