+++ src/utility.ts
@@ -113,22 +113,6 @@ export function getScriptContext(file: ts.SourceFile, seen = new Set<string>()):
 	}
 }
 
-export function red(text: string) {
-	return `\x1b[31m${text}\x1b[0m`;
-}
-
-export function yellow(text: string) {
-	return `\x1b[33m${text}\x1b[0m`;
-}
-
-export function bold(text: string) {
-	return `\x1b[1m${text}\x1b[0m`;
-}
-
-export function suggest(text: string) {
-	return `...\t${yellow(text)}`;
-}
-
 export function isIdentifierWhoseDefinitionMatchesNode(
 	node: ts.Node<ts.ts.Node>,
 	potentialDefinition: ts.Identifier,
